pyproject.toml
[tool.poetry]
name = "proto-boy"
version = "0.1.0"
description = ""
authors = ["Bietola <dincio.montesi@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fire = "^0.4.0"
python-telegram-bot = "^13.11"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
